Refactor: move all remaingin utils to core
