+++ nitorch/vb/mixtures.py
@@ -10,8 +10,7 @@ import math
 from timeit import default_timer as timer
 import matplotlib.pyplot as plt
 from ..core.optim import get_gain, plot_convergence
-from ..utils import softmax
-from ..core.math import besseli
+from ..core.math import besseli, softmax
 import torch
 
 torch.backends.cudnn.benchmark = True
